Fix multiselect for contact lists
